Fixed unit test execution.
There was a subtle bug in the testParsing function that was causing the same unit test
to be run for every expected pair. The key and pairs variables were declared in the wrong
scope so they were being set to the last pair in the set and that last pair was being
tested multiple times.

Fixing this bug also uncovered two failing test cases, which were also fixed.