Correctly handle ISO date values.
This fixes #30. Dates in ISO format were incorrectly being
truncated before the first colon character. As mentioned in #30,
a string 2013-06-06T10:53:59.763Z would be truncated to
2013-06-06T10 before being handed off to the converter function.

Although the code was trying to specify that the string be
split into only two parts by calling String.split(':', 2), this
does not have the desired effect. The string is still split
on all instances of the separator, and only the output array
is truncated to the value of limit. See

http://www.ecma-international.org/ecma-262/5.1/#sec-15.5.4.14